Always try to read distribution specific vars
Let's prevent having undefined variables by declaring them as empty
variables within `default/`. These have the lowest priority in the
variable precedence so are overwritten as soon as the variable is found
elsewhere.

Also increment required Ansible version to 2.9 because of the syntax
used to ignore errors for the first_found lookup plugin.